// data/data.interface.ts
export interface ExploreWorld {
  id: string;
  imgUrl: string;
  title: string;
}

export interface NewFeature {
  imgUrl: string;
  title: string;
  subtitle: string;
}

export interface Insight {
  imgUrl: string;
  title: string;
  subtitle: string;
}

export interface Social {
  name: string;
  url: string;
}